requirements.txt
pandas==2.2.3
plotly==5.24.1
PyYAML~=6.0.2
PyYAML==6.0.2
Requests==2.32.3
openpyxl~=3.1.5
matplotlib~=3.10.0
scikit-learn~=1.6.1